requirements.txt
APScheduler==3.9.1
connexion==2.14.1
PyYAML==6.0
requests==2.25.1
SQLAlchemy==1.4.41
flask-cors==3.0.10